Separate SwipeDetectionDemoViewController to its own class file
